% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_logo_text.R
\name{add_text_logo}
\alias{add_text_logo}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
add_text_logo(fig)
}
\arguments{
\item{fig}{the ggplot item to apply the text logo to}
}
\description{
Add the Text Logo to the ggplot2 figure
}
