% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_mc_base}
\alias{palette_mc_base}
\title{Colors for Moses Cone}
\format{An object of class \code{character} of length 1.}
\usage{
palette_mc_base
}
\description{
Colors for Moses Cone
}
\keyword{datasets}
