% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cone_defaults.R
\name{set_cone_defaults}
\alias{set_cone_defaults}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
set_cone_defaults(style = "print")
}
\arguments{
\item{style}{The default theme style for the R session. "print" or "map".}
}
\description{
\code{set_urbn_defaults} provides a \link{ggplot2} theme formatted according to the
Urban Institute style guide, with sensible defaults.
}
