% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_urbn_main}
\alias{palette_urbn_main}
\alias{palette_mc_base}
\title{Urban Institute palettes updated with Moses Cone Colors}
\format{An object of class \code{character} of length 8.}
\source{
Urban Institute Data Visualization Style Guide
 \url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\usage{
palette_urbn_main

palette_mc_base
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the Urban Institute Data Visualization Style Guide.
\url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\concept{urbn palettes}
\keyword{datasets}
